refactor: Admin dashboard cleanup
