+++ frontend/src/lib/features/round-generator/steps/2-general-data/validation.ts
