+++ frontend/src/lib/stores/flow/FlowStore.ts
@@ -56,11 +56,16 @@ const contractData = {
 	}
 };
 
-export const user = writable<CurrentUserObject | { loggedIn: false }>({ loggedIn: false });
+export const user = writable<CurrentUserObject | { loggedIn: false; addr: null }>({
+	loggedIn: false,
+	addr: null
+});
 export const profile = writable(null);
 // export const transactionStatus = writable({});
 // export const transactionInProgress = writable(false);
-export const addresses = {
+export const addresses: {
+	[key: string]: string;
+} = {
 	NonFungibleToken: contractData.NonFungibleToken[network],
 	MetadataViews: contractData.MetadataViews[network],
 	FungibleTokenMetadataViews: contractData.FungibleTokenMetadataViews[network],
